// Frontend/Angular/src/app/clients/teacher/model/assignmentDetailsDto.ts
/**
 * Teacher api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentDto } from './attachmentDto';

export interface AssignmentDetailsDto { 
    attachments?: Array<AttachmentDto>;
    id?: number;
    subjectId: number;
    subjectName?: string;
    name: string;
    description?: string;
    maxScore?: number;
    allowAfterDeadline?: boolean;
    fineAfterDeadline?: number;
    isHidden?: boolean;
    start?: Date;
    deadLine?: Date;
}